core: refactor how internal symbols are exposed
Use a global symbol in the global scope rather than a normal variable.
Also don't expose that in the tjs namespace.

+++ src/tjs.h
@@ -33,7 +33,7 @@
 typedef struct TJSRuntime TJSRuntime;
 
 typedef struct TJSRunOptions {
-    size_t mem_limit;
+    int mem_limit;
     size_t stack_size;
 } TJSRunOptions;
